+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Migrated repository to `williamhjcho/i18next`
 * Reduce description size
 * Change the namespaces type from `Map<String, Map<String, Object>> -> Map<String, Object>`
+* Adds I18Next.of(BuildContext) from Localizations
 
 ## [0.0.1+5]